Move dialog for qualified recruitment cards into it's own component
